Read decls from resource files

+++ neo/ztech/renderer/frame_data.zig
@@ -67,8 +67,8 @@ const FRAME_MEMORY_ALIGNMENT: usize = 16;
 const CACHE_LINE_SIZE: usize = 128;
 
 // SMP = Symmetric multiprocessing / shared-memory multiprocessing
-var smp_frame: usize = 0;
-pub const NUM_FRAME_DATA: usize = 3;
+var smp_frame: u32 = 0;
+pub const NUM_FRAME_DATA: u32 = 3;
 var smp_frame_data: [NUM_FRAME_DATA]FrameData = undefined;
 var buffer_allocators: [NUM_FRAME_DATA]std.heap.FixedBufferAllocator = undefined;
